cleaned up server routes for reviews
